Add html scaffolding
I've stepped through the file adding all the elements of the basic
layout that we see in the example image.

So 6 main sections, Header, Hero, Info, Quote, Action, Footer, with
further divs and elements inside each section to hopefully make adding
the flexboxes and styling a smooth process.

I think I need to revisit the image element in the hero section to make
sure it is an actual img, with alt text providing the placeholder, not
just plain text in a div. And some other todos are stripping out the
list bullets and link underlining.