Add colours and font styling
As we're given free choice of what order to tackle the page design in,
I've opted to split the CSS styling up into two sections, visual styling
and then flexboxes and positioning.

This first commit therefore just concerns the visual styling, as in
colour choices and sizing and styling the text throughout the page.
Example image 2 was used to guide the process, with comparisons to image
1 to make sure I was achieving the right appearance.

I've split the initial sections of my css file out into colour choices,
font sizing, font styling, just for ease of reference if I need to go
back and change things. So far it all looks to be matching the requested
design though, now to tackle the positioning and flex aspects!