Add missed styling
When filling out the readme to update my progress I realised I had
missed several styling elements while working on the CSS.

This commit therefore adds the grey background to the image placeholder
in the hero section. It adds a solid white border to the button in the
call to action section, and removes any border from the button in the
hero section. It also adds rounded corners to the elements that were
supposed to have them.

And lastly I introduced max widths to the containers in the different
sections. So 1000px for those in the header, hero section and call to
action. 825px in the info section and quote section. It's a reasonable
starting point for ensuring the page elements will match what is given
in the example.